requirements.txt
faiss_cpu==1.10.0
numpy==2.2.3
openai==1.63.2
python-dotenv==1.0.1
